pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "minio-uploader"
version = "0.1.0"
description = "CLI for working with MinIO via S3"
authors = [{name="Bogdan"}]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "boto3",
    "python-dotenv",
]

[tool.setuptools.packages.find]
where = ["src"]

[project.scripts]
minio-uploader = "minio_uploader.cli:main"
